+++ ivy/functional/backends/jax/experimental/layers.py
@@ -1,10 +1,14 @@
+# global
 from typing import Optional, Union, Tuple, Literal
-import ivy
-from ivy.functional.backends.jax import JaxArray
 import jax
 import jax.lax as jlax
 import jax.numpy as jnp
 import math
+
+# local
+import ivy
+from ivy.functional.backends.jax import JaxArray
+from ivy.functional.backends.jax.random import RNG
 
 
 def general_pool(inputs, init, reduce_fn, window_shape, strides, padding):
@@ -318,3 +322,28 @@ def fft(
     if norm != "backward" and norm != "ortho" and norm != "forward":
         raise ivy.exceptions.IvyError(f"Unrecognized normalization mode {norm}")
     return jnp.fft(x, n, dim, norm)
+
+
+def dropout1d(
+    x: JaxArray,
+    prob: float,
+    /,
+    *,
+    training: bool = True,
+    data_format: str = "NWC",
+    out: Optional[JaxArray] = None,
+) -> JaxArray:
+    if training:
+        if data_format == "NWC":
+            perm = (0, 2, 1) if len(x.shape) == 3 else (1, 0)
+            x = jnp.transpose(x, perm)
+        noise_shape = list(x.shape)
+        noise_shape[-1] = 1
+        _, rng_input = jax.random.split(RNG.key)
+        mask = jax.random.bernoulli(rng_input, 1 - prob, noise_shape)
+        res = jnp.where(mask, x / (1 - prob), 0)
+        if data_format == "NWC":
+            res = jnp.transpose(res, perm)
+        return res
+    else:
+        return x
