+++ ivy/functional/backends/tensorflow/experimental/layers.py
@@ -216,3 +216,26 @@ def fft(
         ret = tf.signal.fft(x, operation_name)
     ret = _fft_norm(ret, dim, norm)
     return ret
+
+
+def dropout1d(
+    x: Union[tf.Tensor, tf.Variable],
+    prob: float,
+    /,
+    *,
+    training: bool = True,
+    data_format: str = "NWC",
+    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
+) -> Union[tf.Tensor, tf.Variable]:
+    if training:
+        if data_format == "NCW":
+            perm = (0, 2, 1) if len(x.shape) == 3 else (1, 0)
+            x = tf.transpose(x, perm)
+        noise_shape = list(x.shape)
+        noise_shape[-2] = 1
+        res = tf.nn.dropout(x, prob, noise_shape=noise_shape)
+        if data_format == "NCW":
+            res = tf.transpose(res, perm)
+        return res
+    else:
+        return x
