+++ ivy/functional/backends/tensorflow/layers.py
@@ -433,29 +433,3 @@ def conv_general_transpose(
     if data_format == "channel_first":
         res = tf.transpose(res, (0, dims + 1, *range(1, dims + 1)))
     return res
-
-
-conv_general_transpose.unsupported_dtypes = ("bfloat16",)
-
-
-def dropout1d(
-    x: Union[tf.Tensor, tf.Variable],
-    prob: float,
-    /,
-    *,
-    training: bool = True,
-    data_format: str = "NWC",
-    out: Optional[Union[tf.Tensor, tf.Variable]] = None,
-) -> Union[tf.Tensor, tf.Variable]:
-    if training:
-        if data_format == "NCW":
-            perm = (0, 2, 1) if len(x.shape) == 3 else (1, 0)
-            x = tf.transpose(x, perm)
-        noise_shape = list(x.shape)
-        noise_shape[-2] = 1
-        res = tf.nn.dropout(x, prob, noise_shape=noise_shape)
-        if data_format == "NCW":
-            res = tf.transpose(res, perm)
-        return res
-    else:
-        return x
