+++ ivy/functional/backends/torch/experimental/layers.py
@@ -409,3 +409,28 @@ def fft(
     if norm != "backward" and norm != "ortho" and norm != "forward":
         raise ivy.exceptions.IvyError(f"Unrecognized normalization mode {norm}")
     return torch.fft.fft(x, n, dim, norm, out=out)
+
+
+def dropout1d(
+    x: torch.Tensor,
+    prob: float,
+    /,
+    *,
+    training: bool = True,
+    data_format: str = "NWC",
+    out: Optional[torch.Tensor] = None,
+) -> torch.Tensor:
+    if training:
+        if data_format == "NWC":
+            perm = (0, 2, 1) if len(x.shape) == 3 else (1, 0)
+            x = torch.permute(x, perm)
+        # ToDo: switch to native dropout1d once torch version is updated.
+        noise_shape = list(x.shape)
+        noise_shape[-1] = 1
+        mask = torch.rand(noise_shape) > prob
+        res = torch.where(mask, x / (1 - prob), torch.zeros_like(x))
+        if data_format == "NWC":
+            res = torch.permute(res, perm)
+        return res
+    else:
+        return x
